test(footer): migrate basic
